requirements.txt
python-dotenv==1.0.0
requests==2.31.0
elevenlabs
moviepy==1.0.3
tqdm==4.66.1
pyyaml==6.0.1
deepl==1.16.0
google-cloud-translate==3.11.3
openai
